// src/app/admin/components/types.ts
export class User {


	 _id: String;
    username: String;
    email : String;
    password: String;
    image: String;
    role: String;
    createdtime: String
}


export class Testimonal {


	 _id: String;
    author: String;
    content: String;
    image: String;
    description: String;
    createdtime: String

}


export class Ebook {


	 _id: String;
    author: String;
    title : String;
    content: String;
    image: String;
    description: String;
    category: String;
    tags: String;
    createdtime: String
}

export class Blog {


	 _id: String;
    author: String;
    title : String;
    content: String;
    image: String;
    description: String;
    category: String;
    tags: [String];
    createdtime: String
}


export class EmbedVideo {


	 _id: String;
    author: String;
    title : String;
    content: String;
    image: String;
    description: String;
    category: String;
    tags: String;
    createdtime: String
}



export class PPT {


	 _id: String;
    author: String;
    title : String;
    content: String;
    image: String;
    description: String;
    category: String;
    tags: [String];
    createdtime: String
}


export class Template {


	 _id: String;
    author: String;
    title : String;
    content: String;
    image: String;
    description: String;
    category: String;
    tags: [String];
    createdtime: String
}